+++ core/cmake/get_python_compile_settings.py
@@ -3,7 +3,7 @@
 
 #   BAREOS(R) - Backup Archiving REcovery Open Sourced
 #
-#   Copyright (C) 2020-2020 Bareos GmbH & Co. KG
+#   Copyright (C) 2020-2024 Bareos GmbH & Co. KG
 #
 #   This program is Free Software; you can redistribute it and/or
 #   modify it under the terms of version three of the GNU Affero General Public
@@ -34,7 +34,7 @@ except:
 for var in ("CC", "BLDSHARED"):
     value = sysconfig.get_config_var(var)
     print(
-        'message(STATUS "Python{0}_{1}\ is\  {2}")'.format(
+        'message(STATUS "Python{0}_{1} is  {2}")'.format(
             sys.version_info[0], var, value
         )
     )
@@ -47,7 +47,7 @@ for var in ("CC", "BLDSHARED"):
         value = ""
     # as these vars contain the compiler itself, we remove the first word and return it as _FLAGS
     print(
-        'message(STATUS "Python{0}_{1}_FLAGS\ is\  {2}")'.format(
+        'message(STATUS "Python{0}_{1}_FLAGS is  {2}")'.format(
             sys.version_info[0], var, value
         )
     )
@@ -56,7 +56,7 @@ for var in ("CC", "BLDSHARED"):
 for var in ("CFLAGS", "CCSHARED", "INCLUDEPY", "LDFLAGS"):
     value = sysconfig.get_config_var(var)
     print(
-        'message(STATUS "Python{0}_{1}\ is\  {2}")'.format(
+        'message(STATUS "Python{0}_{1} is  {2}")'.format(
             sys.version_info[0], var, value
         )
     )
@@ -67,7 +67,7 @@ for var in ("EXT_SUFFIX",):
     if value is None:
         value = ""
     print(
-        'message(STATUS "Python{0}_{1}\ is\  {2}")'.format(
+        'message(STATUS "Python{0}_{1} is  {2}")'.format(
             sys.version_info[0], var, value
         )
     )